world-space-main/requirements.txt
asgiref==3.5.2
boto3==1.26.30
botocore==1.29.30
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
decorator==5.1.1
distlib==0.3.6
Django==3.2.16
django-cors-headers==3.13.0
django-localflavor==3.1
django-phone-field==1.8.1
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
dynaconf==3.1.11
filelock==3.8.2
hvac==1.0.2
idna==3.4
importlib-metadata==5.1.0
Jinja2==3.1.2
jmespath==1.0.1
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
Pillow==9.3.0
platformdirs==2.6.0
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
python-stdnum==1.18
pytz==2022.6
requests==2.28.1
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.13
validators==0.20.0
virtualenv==20.17.1
win32-setctime==1.1.0
zipp==3.11.0
